Deep debt: NaN-safe optimizers, production unwrap elimination, num_cpus→std evolution, 100+ new tests
- Evolve all partial_cmp().unwrap() in optimize/ to NaN-safe unwrap_or(Ordering::Equal)
  (nelder_mead.rs, solver_state.rs, multi_start.rs — 7 call sites)
- Replace ESN::predict .unwrap() with Result error for untrained state
- Replace SNN Dense layer .unwrap() with Result error for uninitialized weights
- Replace scheduler .expect() with .unwrap_or_else() fallback
- Evolve num_cpus FFI dependency to std::thread::available_parallelism() across
  barracuda (3 files), toadstool (2 files), config (2 files), server (6 files)
- Remove num_cpus from barracuda Cargo.toml (zero FFI for CPU count)
- Add 65+ new unit tests: byob_types (16), jobs (8), requests (9), auth (23),
  agents (13), graph_types (20), capabilities (14)
- Total core tests: 3,667 (barracuda 1237, toadstool 1040, common 674,
  config 316, server 400)
- Zero clippy warnings, zero build warnings across workspace

+++ crates/runtime/gpu/tests/unified_memory_e2e_tests.rs
@@ -148,10 +148,7 @@ async fn test_e2e_large_data_transfer() {
     // Write in chunks
     let chunk_size = 1024 * 1024; // 1MB chunks
     for (i, chunk) in write_data.chunks(chunk_size).enumerate() {
-        buffer
-            .write_async(i * chunk_size as usize, chunk)
-            .await
-            .unwrap();
+        buffer.write_async(i * chunk_size, chunk).await.unwrap();
     }
 
     // Read back and verify
